added commands module with some tests added stub db and optional constructor db arg
